mm61xx: (APP-3244) ignore iw reg set on running driver
Our support for this is a bit half-baked and causes numerous issues
(see linked tickets).

For now, let's just make the HaLow regdom entirely separate
from the global regdom by ignoring any notifications.
Note that this looks pretty safe: we're already ignoring
notifications in some circumstances (e.g. when we have an AP
configured or when OTP bits are set).


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn