persistent-vars-storage: (PR #752): Add READ -a flag to read all vars
Define all the known vars and a flag that prints them all. This is only
useful for developers who forget the names of vars.

Missing vars are printed like `mm_mode=`, so there is no way to
distinguish empty and missing vars.  Non printable characters are not
escaped, nor are newlines or = signs.  This is only useful for
developers.

This works on EKH01, although it is not very useful. It spews a lot of
text to stderr.

Approved-by: James Haggerty
Approved-by: Sophronia Koilpillai