wizard-config: (APP-3299, PR #751) auto-configure client device
This is like a cut-rate DPP: requires TCP/IP connection, and
no security on the link.

Most of the client/server side is run via a ucode script (wizard-config).
Refer to HELP at the top of this for more info about what is going
on here.

This is to get around the problem of wizards on bridged station devices
which we either don't have access to _or_ the problem of wizards which
change IPs under you.

Client config will only happen on first boot after a reset.

Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn