luci-mod-home: (APP-3322, PR #762) add encryption options to uplink connection
When I made this, it was as minimal as possible. Basically only supported SAE.
This is a bit too minimal for 2.4, when you might have different
options depending on what device you're connecting to.

This becomes more important for the new wizard because when you set
a 2.4 uplink the expectation is you go to the homepage or quickconfig
to 'fix' the connection.

Look, this is pretty terrible. It's basically just a minimal
amount of code copied out of the quick config page, and not enough
to support all encryption modes. However, sharing this code well is tricky,
and I don't really want to bite this off at the moment.


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn