mm61xx: (APP-3349, PR #781) better handle healthcheck fails
Artini seems to have some flakiness in the comms over the SDIO bus,
causing it to occasionally fail healthchecks with checksum
errors.

(in an overnight test, 177 / 11586450)

Since this doesn't indicate that the chip is actually dead, it's
useful to be really _sure_ before a healthcheck leads to the chip
dying.


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn