morseupgrade: APP-3260: (PR #780): Add manual firmware upload button
Add a button to manually upload a firmware file like on the flash.js
page. Copy the code from there. Manually uploading writes to the same
file (sysupgrade.bin) as the automatic upgrade, which is different to
the flash.js page (firmware.bin).

Pressing the automatic button hides the manual button to avoid confusion
during the automatic process, and dodge dealing with any asynchronous
issues from two processes writing to the same file.  Pressing the manual
button does not hide the automatic button, because there is a ui popup
during the entire process.

Approved-by: Sophronia Koilpillai
Approved-by: Arien Judge
Approved-by: James Haggerty