luci-app-morseconfig/luci-mod-home: (APP-3356) use zone name in the actual uci config
Before we were using default configs and this was working because we gave
the zones the same section name as the name in the uci config.

i.e.

config zone foo
	option name foo

But the default zones do not have section names.

i.e.

config zone
    option name foo

This also led to colours being wrong.

Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn