luci-app-morseconfig: (APP-3362, PR #793) having WDS option for 2.4 was a footgun
This would only work if the thing you're connected to was
a 2.4 AP with a compatible WDS implementation. Given this is something
well outside our control, it's more likely the user will incorrectly
select this, so let's remove the footgun from the quick config.

It will still be possible to set WDS via the advanced config
(and currently this page does _not_ attempt to change it, which could
lead to some confusion, BUT at least the user messed themselves
up by going to advanced config...).


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn