hostapd_s1g/wpa_supplicant_s1g: (APP-3347, PR #802) fix dpp channel patch typo
I did this exact sequence of stupid last time.

What happened is that I pulled the patch off the top of my hostapd
tree forgetting that I'd had to amend it last time, so duplicated
this error.

Then I initially didn't _see_ the error because when
hostapd_s1g is compiled it doesn't compile any of the wpa_supplicant
stuff, so we only saw it when I actually put the other patch in.


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn