wpa_supplicant_s1g: (APP-3372, PR #800) make wpa_supplicant dbus optional
The intent here is to make it optional in the Makefile.

If we want to have it always on for ekh01, I'm fine with this, and we
can set it always enabled in the board config (and indeed matter
could have this as a dependency). However, I would prefer NOT to ship
this on Artini; it also blows out our 16mb image.


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn