luci-app-morseapwizard: (APP-3375, PR #806) keep 2.4 AP for config
Keep 2.4 AP on LAN in all situations so people can more easily use their mobile device to configure.

We also can’t think of a valid use case where you’d want to bridge this; the main reason to do it in the first case was to keep the APs aligned.

Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn