luci-mod-home: (APP-3381) make uplink connection tick/cross linkable
Previously, the only way to expand an uplink card was to click
the expand up the top. Make it so that if there's a maxContents
on any card, the click-to-expand is enabled and hover behaviour
is setup, and if there's no maxContents nothing happens
(this also affects the prplmesh card when it has no agents).


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn