luci-mod-home: (APP-3386, PR #816) simplify refocus code
When we move a DOM element around (as we do here by
retaining a bit of the old DOM rather that rerender) to maintain
the user's position we need to call .focus().

For some reason, I got a bit caught up in the particular case
(wanting to refocus on an input box and have the cursor in the right
position), but this is not necessary.


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn