morse-bundle: APP-3401: (PR #825) Unspecify the ttyd listen interface
ttyd listens only on @lan by default, meaning @wlan would not work for
a wizard using user. Make it listen everywhere, which should be OK as
that is what dropbear does. We are relying on the wan firewall to
prevent unexpected access.

Tested by connecting to a AP dut via an extender, testing ttyd, then
modifying the AP to the bridged mode and testing ttyd.


Approved-by: James Haggerty
Approved-by: Sophronia Koilpillai