luci-app-morseconfig: (APP-3062, PR #832) fix moving icons
The icons when clicked were jumping around as the spinner
was introduced. Particularly obvious on mobile.

(the tricky thing here is that we're using CSS to
hide the normal button and replace it with an icon,
and we also need to handle the '.spinning' class
which is used to show progress)


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn