morse-modeswitch-button: add Morse Micro modeswitch button package
This package adds Artini button functionality to OpenWrt.
The Artini button is used for dpp push button, factory reset, and
on longer presses, includes a mode change with that factory reset.

If pressed and release quickly, dpp starts.
If pressed and held until the system LED flashes slowly, move to
AP mode and factory reset.
If pressed and held until the system LED flashes quickly, move to
station mode and factory reset.