wpa_event_listener: (APP-3072, PR #849) if hostapd/wpa_supplicant restarts, clear LED
If hostapd/wpa_supplicant restarts, then wpa_event_listener
quits and any timeout is lost. But it doesn't call the action
script, which means the LED is stuck flashing with the wrong colour.

The easy fix here is to call it finished on termination.
This works fine for hostapd (since hostapd will clear the
cached keys, so it's a proper termination and it's fine
to initiate a dpp push button again), but for wpa_supplicant
it's not great since this will let the user start another
DPP session and potentially clash on the hostapd side.
However, I think this is an edge case, and would prefer
to keep the logic simple and make sure wpa_event_listener
properly cleans up when it exits.


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn