luci-app-morseconfig: (APP-3527, PR #871) correct discovery of interface name
This looks like it was always wrong (but how?). The good news
is it only affects two pages (support and stats), and since we're pretty consistently
coming up as the default (wlan0) it's only a big issue on Artini
when prplmesh is involved (since even mesh uses wlan0 for the AP).


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn