luci-app-ekhwizards/uci-app-morseapwizard: (APP-3380, PR #878) make it possible for wizards to co-exist
Previously, the ekhwizards would conditionally create an alias depending
on the existence of the other package. Unfortunately, this didn't work
correctly, probably due to the act of loading the json replacing
the other wizard's JSON entry.

This is the other dodgy alternative: have both packages
create an alias for selectwizard, and we don't really
care which one wins (since the only use case for
installing multiple wizards is for testing).


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn