wizard-config: (APP-3689) iterate over all server addresses with wizard-config on failure
We were just trying the first address found (or if it was a DHCP client,
preferring the gateway address).

This mostly seemed to work, but it's quite dangerous: umdns advertises _all_ the
available IPs, so it will cheerfully advertise IPs that are not on the
subnet you're connected to. e.g.

	"_wizard-config._tcp": {
		"halowlink1-1a27": {
			"iface": "br-lan",
			"host": "halowlink1-1a27.local",
			"port": 80,
			"txt": [
				"path_v1=/wizard-config.json"
			],
			"ipv4": [
				"192.168.12.1",
				"10.22.121.111",
				"10.11.0.131"
			],
			"ipv6": [
				"fd3c:afb7:5bac::1",
				"fe80::9683:c4ff:fe57:1a28",
				"fe80::9683:c4ff:fe57:1a27"
			]
		}
	}

This is a AP only (no router) situation, but it advertises all of the
management IP, the static IP, and the DHCP client IP.

While I was here, I removed the untested IPv6 support, since it didn't
see to be working for me anyway and I didn't want it to slow down
the logic now we iterate throu…

Approved-by: Lyall Beveridge
Approved-by: Evan Benn