netifd-morse: (APP-3679, PR #915) make sure wlan0 is never 'stolen' by 2.4
Previously, if a non-morse device had wlan0, we were waiting until it
was cleaned up by netifd when it setup the device (at which point
wlan0 would be remove and we could grab it).  Unfortunately, when the
device was completely disabled it wouldn't happen (i.e.  confusingly,
if you disabled 2.4, your morse device would end up on wlan1).

A better approach here is to simply put an early hotplug to remove the
AUTOVIF interface created by the kernel, leaving it to netifd in all
cases.

It's useful for use to have a fixed first morse
wifi interface for two reasons:
 - morse_cli has a default that usually works
 - our LED configuration is tied to this rather than to the phy,
   since the phy changes on module load/unload (which we require for
   certain changes)

Approved-by: Lyall Beveridge
Approved-by: Evan Benn