luci-mod-home: (APP-3677, PR #909) prplmesh topology view - be more resilient
The prplmesh topology view integrates information from different sources:

- devices associated with APs (incl signal strength)
- those reported as mesh neighbours

Sometimes, devices can be listed as associated but they're a neighbour
of another node (which leads to duplicate links). Sometimes,
a device can be neighbour but not appear in the devices list
at all. This PR tries to adjust to more of this by:

- not showing AP->STA links if it's in the devices list (i.e. a mesh thing)
- if it's listed as a neighbour but isn't in the devices list, make
  it an agent (with limited info)

While here, I've made a small clean-up to the signal strength display
to avoid showing 'Unknowndbm'.


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn