luci-*: (APP-3263, PR #938) cache bust qrcode svg
Without this, it was possible for a cache qrcode from another
device (or version) to appear. Using the sessionid means we still
get caching once the user's logged in, but we're guaranteed a
fresh request for every session. We chop off some of the sessionid
so we're not sending what is effectively our auth token as part
of the header...

Note that the best way to cache bust would be to have a unique
device id, but this is not currently exposed by luci.
We can't use the usual version based cache busting since this
is specific to a build, not to a device. Ideally, we might just
encode the MAC address into the dpp_qrcode.svg filename, but
then we'd have to figure out a way to feed the halow MAC
into the frontend. So, it just doesn't seem worth the extra
complexity/effort at this time.


Approved-by: Evan Benn
Approved-by: Lyall Beveridge