APP-3580 (PR #936): netifd/morse.sh: Trigger smart_manager service restart
When morse.sh reconfigures the halow interfaces, after informing procd and
starting hostapd/wpa_supplicant, restart the smart_manager service. Stop the
service when tearing down the ifaces. smart_manager needs to start after
hostapd, as it communicates with hostapd and the kernel. Before this fix,
smart_manager would be restarted quickly and communicate with the 'old' hostapd
(that had not been restarted yet).

The common bug was that smart_manager would switch
the AP back to the old channel bandwidth if you tried to reconfigure
the channel bandwidth.

With this patch, smart_manager still starts up before hostapd is ready,
however it already contains logic to wait for hostapd to be in the
correct state. The prior implementation here was bypassing that
waiting.

I tested this by:
- rebooting: We now disable smart manager at boot
- wifi down
- wifi up
- service smart_manager reload
- service smart_manager restart
- reconfigure smart_manager
- reconfigure halow channel bandwidth
- reconfigure halow AND smart_manager: There is a race condition, a potential
double restart of smart_manager. It seems benign other than potential log failure messages.

Approved-by: Sophronia Koilpillai
Approved-by: James Haggerty