luci-app-morseconfig: (APP-3685, #941) don't right align action on stats page
Because we're using a table here which is now mostly centre aligned with
Argon, and there are only two columns, right aligning the actions looks
weird.

Vaguely tempted to just rework the page, but for now let's just centre
align the actions by not pulling in the usual action styling.


Approved-by: Lyall Beveridge
Approved-by: Evan Benn