luci-app-ekhwizards: (APP-3245, PR #952) force wizards to lose info when changing wizard
This was sort of ok, but potentially confusing, and meant that
a particular mode was sometimes preselected rather than forcing the user
to choose. It also disagreed with the comments.

I think moving to and from mesh is enough of a change that we should
force people to go through each of the selections.


Approved-by: Lyall Beveridge
Approved-by: Evan Benn