luci-app-ekhwizards: (APP-3245) use default AP ssid even if wizard not run
We were aggressively clearing the AP mode if the wizard hadn't been run
to force the user to select AP/Client in the wizard.

Unfortunately, later we relied on this mode when loading the ssid/key,
which meant that if the user had:
- modified the SSID/key before running the wizard at all
- then run the wizard

It would discard their custom SSID/key and present them with the default
(unexpected!).


Approved-by: Evan Benn
Approved-by: Milad Mohtashamirad