luci-app-morseconfig: (APP-3751, PR #974) add defaults to new wifi-iface row
When you added a new wifi-iface row on the quick config page, it
was empty and everything had to be added. This makes it harder
for users to have a sane setup, particularly if they've removed
their default iface.

Since we were patching handleAdd anyway, I've also changed it
do it doesn't do a (surprise) save to achieve a re-render, and
simply loads the page from the JS uci cache.


Approved-by: Lyall Beveridge
Approved-by: Evan Benn