luci-app-ekhwizards: (APP-3374, PR #1003) ensure that masq/mtu_fix is correctly set on lan
When we moved to using the default OpenWrt setup, we never created our
'lan' zone and hence never added 'masq=1'. When using our standard 'routed'
AP setup - where ahwlan -> lan - this meant that masq wasn't set.

To resolve this, we:
- reset mtu_fix/masq in the wizard
- create mtu_fix/masq in forwarding

For consistency/clarity, we also now ensure that the forwarding is as we expect
in ensureNetworkExists based on local. While I was here, I also made sure that
if the network was _not_ local, it was removed from umdns/camera-onvif-server.


Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge