luci-app-ekhwizards: (APP-3775, PR #1004) Mesh AP should default to SAE encryption
Now that the Mesh AP has a default Wi-Fi key, it was possible
to go through the wizard without ever causing a write of the key,
which we relied on to set the encryption type.

Instead, let's just rely on setting the encryption separately
(note that the encryption for the mesh interface is set
in parseWizardOptions, so I've also removed that code).


Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge