luci-app-morseconfig: (APP-3789, PR #1014) make sure there's always a WAN
Because our quickconfig page doesn't allow users to add network
interfaces at the moment, that the default ekh devices only have LAN
makes it hard to do anything useful on this page before running the wizards.

Since OpenWrt installs wan firewall rules by default, it makes sense to
have a WAN interface even if it doesn't have any devices.

I initially added this to the board config in morse_openwrt - which allowed
adding ahwlan/wlan as well for ekh/artini - but unfortunately this would have
required a patch to the main part of OpenWrt to let it create network
interfaces without any ports/devices, and I decided the lowest impact
change (since it’s quickconfig that ‘wants’ this) would be to put it as a
uci default in the quick config package.


Approved-by: Andrew Dodd
Approved-by: Lyall Beveridge