luci: (APP-3643, PR #1021) Update SSID and Security Key fields to be disabled when not relevant
Motivation: In some modes (such as monitor mode), the SSID and Security
Key fields are not necessary, but they exist in the UI which can cause
confusion.

This change updates the fields to become disabled with the placeholder
text "Not required" if the field itself is not required. For the SSID
field, this is when the wirelss mode is `none` or `monitor`. For the
security key field, this is when the encryption mode is `none`.


Approved-by: James Haggerty
Approved-by: Evan Benn