luci-app-*: (APP-3788, PR #1020) make sure wan firewall is always in a good state
Previously, we were not attempting to set the WAN firewall rules,
which implicitly relied on the wan firewall being correctly configured.
Since the intent of the wizard code is that it should adjust to almost
any existing configuration - and the code in ensureNetworkInterface
seemed to imply it would correctly setup a firewall - we copy
the default wan rules into here.

The downside is that the wizard is now destructive of all aspects of
your firewall setup.


Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge