morse-bundle/morse-modeswitch-button: (APP-3771, PR #1029) be more resilient to slow hotplug
Defend against a slow hotplug meaning we never get the chance
to set our custom wifi defaults.

We want this order:
- hotplug -> wifi config -> /lib/wifi/morse.sh
- uci-defaults/30_uboot-envtools (so persistent_vars_storage.sh works)
- uci-defaults/98-morse-wireless-defaults

See /etc/init.d/boot for how this is triggered (kmodloader -> uci-defaults).

But because, well, it's a hotplug, it's hard to nicely know when it's
finished at boot time.

Really, I think OpenWrt should have a mechanism to easily hook into
the hotplug after the wifi-device/wifi-iface creation (e.g.  call a
uci-defaults style script with these two as arguments after each add,
or I guess one could attach a procd script for this). However, in
this case - though it would be easy to add - we would still have
the dependency on 30_uboot-envtools, so I've gone for the terrible
hack.


Approved-by: Evan Benn
Approved-by: Andrew Dodd
Approved-by: Lyall Beveridge