luci-lib-morseui (APP-3848, PR #1030) Ensure monitor and none mode can be saved
Motivation: The changes from commit d2ee6ff3 made it impossible to save
and apply the config to set `none` or `monitor` mode. This change fixes
the bug to allow saving.

Cause: The changes in d2ee6ff3 made the SSID input disabled and have no
value for the `none` and `monitor` modes. However, because the SSID
field is not optional and the value is empty there is no way to save the
form without encountering the empty field validation. This change works
by overriding the isActive definition for the SSID field to also
consider the field inactive if it is not required (instead of just when
it is "hidden"). Looking at the isActive use in Luci this appears to be
safe.

Approved-by: James Haggerty
Approved-by: Lyall Beveridge