luci-mod-home: (APP-3743, PR #1041) make sure dongle appears on card diagram
I somehow missed updating the ethernet ports with the dongle
when I added some basic dongle support to the diagram, so the diagram
was inconsistent between the wizard/config page and the home page
(i.e. the LTE uplink wasn't appearing - though it appears as a 'router'
anyway, so it's not that useful).

Approved-by: Andrew Dodd
Approved-by: Lyall Beveridge