luci-lib-morseui (APP-3919, PR #1043) Switching to client mode always clears password
Motivation: When changing to client mode on the quick config screen, we
want the password to be empty. I.e. we don't want to fill in the
device's default WiFi password, as the default password is probably only
going to be valid to connect to the device itself. By making the field
empty we ensure the user fills a value in.

The UI had a bug which causes it to set the password to be the default
password when transitioning from modes like "None" to a station mode.

This was due to a number of interacting issues, such as:
 - Using a bare `<input>` that did not have an `id` set meant the form
   would fallback to retrieving the saved `_wpa_key` value instead of
   the latest value from the form (i.e. it would ignore the previous
   clobbering to "").
 - The `mode` option onchange handler updates both the `encryption` and
   `_wpa_key` options. The `encryption` option onchange handler also
   sometimes updates the `_wpa_key` option. …

Approved-by: Sophronia Koilpillai
Approved-by: James Haggerty
Approved-by: Lyall Beveridge