luci-app-morseconfig: (APP-4010, PR #1057) correct type for network
luci is a bit loosey-goosey with types, and generally any notional
'array' access can handle it if it's not an array (e.g. using L.toArray;
usually obscured because if the form is defined to expect a list
this will automatically happen).

Nevertheless, we should not create incorrect types ourselves.
Hilariously, this _won't_ affect the testing, because in our
'uci' format files we also can't distinguish between a single
option and a single element list.


Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge