luci-app-morseconfig: (APP-3995, PR #1055) allow saving config if readonly item empty
This is a minimally invasive way to solve this. In some ways, what we should
do for readonly is not run validation at all (and never call write); at the
moment, for instance, since 'encryption' is not set when configuring
easymesh for WPS loading it on the quick config page will persist a 'none'
encryption value.

However, this is a minimally invasive fix for now. If we continue to see
problems, it might be worth fixing in luci.


Approved-by: Sophronia Koilpillai
Approved-by: Andrew Dodd
Approved-by: Evan Benn