luci-app-morseconfig: (APP-3894, PR #1063) reduce user confusion around empty networks
The wizard has a tendency to leave empty networks around, but until you understand the quick config page it's not that obvious which networks are unused.

This makes you pay more attention than you should to networks which are irrelevant.

So we add a warning tooltip to warn them that it's not normal and that
they can make it normal by adding devices/interfaces to it. This is the
better than nothing quick fix.


Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge